move data into route handler
